# client/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    url: https://github.com/moov-io/watchman
  description: Moov Watchman offers download, parse, and search functions over numerous
    U.S. trade sanction lists for complying with regional laws. Also included is a
    web UI and async webhook notification service to initiate processes on remote
    systems.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Watchman API
  version: v1
servers:
- description: Local development
  url: http://localhost:8084
...
paths:
...
  /search:
    get:
      operationId: search
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the system's logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Search across Name, Alt Names, and SDN Address fields for all
          available sanctions lists. Entries may be returned in all response sub-objects.
        explode: true
        in: query
        name: q
        required: false
        schema:
          example: John Doe
          type: string
        style: form
      - description: Name which could correspond to an entry on the SDN, Denied Persons,
          Sectoral Sanctions Identifications, or BIS Entity List sanctions lists.
          Alt names are also searched.
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: Jane Smith
          type: string
        style: form
      - description: Physical address which could correspond to a human on the SDN
          list. Only Address results will be returned.
        explode: true
        in: query
        name: address
        required: false
        schema:
          example: 123 83rd Ave
          type: string
        style: form
      - description: City name as desginated by SDN guidelines. Only Address results
          will be returned.
        explode: true
        in: query
        name: city
        required: false
        schema:
          example: USA
          type: string
        style: form
      - description: State name as desginated by SDN guidelines. Only Address results
          will be returned.
        explode: true
        in: query
        name: state
        required: false
        schema:
          example: USA
          type: string
        style: form
      - description: Providence name as desginated by SDN guidelines. Only Address
          results will be returned.
        explode: true
        in: query
        name: providence
        required: false
        schema:
          example: USA
          type: string
        style: form
      - description: Zip code as desginated by SDN guidelines. Only Address results
          will be returned.
        explode: true
        in: query
        name: zip
        required: false
        schema:
          example: USA
          type: string
        style: form
      - description: Country name as desginated by SDN guidelines. Only Address results
          will be returned.
        explode: true
        in: query
        name: country
        required: false
        schema:
          example: USA
          type: string
        style: form
      - description: Alternate name which could correspond to a human on the SDN list.
          Only Alt name results will be returned.
        explode: true
        in: query
        name: altName
        required: false
        schema:
          example: Jane Smith
          type: string
        style: form
      - description: ID value often found in remarks property of an SDN. Takes the
          form of 'No. NNNNN' as an alphanumeric value.
        explode: true
        in: query
        name: id
        required: false
        schema:
          example: "10517860"
          type: string
        style: form
      - description: Match percentage that search query must obtain for results to
          be returned.
        explode: true
        in: query
        name: minMatch
        required: false
        schema:
          example: 0.95
          format: float
          type: number
        style: form
      - description: Maximum results returned by a search. Results are sorted by their
          match percentage in decending order.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          example: 25
          type: integer
        style: form
      - description: Optional filter to only return SDNs whose type case-insensitively
          matches.
        explode: true
        in: query
        name: sdnType
        required: false
        schema:
          $ref: '#/components/schemas/SdnType'
        style: form
      - description: Optional filter to only return SDNs whose program case-insensitively
          matches.
        explode: true
        in: query
        name: program
        required: false
        schema:
          example: SDGT
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
          description: SDNs returned from a search
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occurred, see response body.
      summary: Search
      tags:
      - Watchman
  /search/us-csl:
    get:
      description: Search the US Consolidated Screening List
      operationId: searchUSCSL
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the system's logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Name which could correspond to an entry on the CSL
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: Jane Smith
          type: string
        style: form
      - description: Maximum number of downloads to return sorted by their timestamp
          in decending order.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          example: 25
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
          description: SDNs returned from a search
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occurred, see response body.
      summary: Search US CSL
      tags:
      - Watchman
  /search/batch:
    post:
      description: Search, in a batch
      operationId: searchBatch
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csvFile:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: SDNs returned from a search
        "400":
          description: Error occurred, see response body.
...
components:
  schemas:
...
    SdnType:
      description: Used for classifying SDNs — typically represents an individual
        or company
      enum:
      - individual
      - entity
      - vessel
      - aircraft
      example: individual
      type: string
...
    Error:
      properties:
        error:
          description: An error message describing the problem intended for humans.
          example: Example error, see description
          type: string
      required:
      - error